"fix" camera bug
