+++ lib/index.js
@@ -19,26 +19,12 @@ var raylib_1 = require("./src/bindings/raylib");
 var colors = require("./src/types/colors");
 var camera = require("./src/types/camera");
 var maths = require("./src/types/maths");
+var models = require("./src/types/models");
+var shaders = require("./src/types/shaders");
 var enums = require("./src/types/enums");
 var misc = require("./src/types/misc");
 var texture = require("./src/types/texture");
-var raylib = __assign({}, raylib_1.default, { 
-    // wrapper fixes
-    /**
-     * UpdateCamera is wrapped to allow object reference.
-     *
-     * @see UpdateCameraWrap()
-     */
-    UpdateCamera: function (camera) {
-        var newCamera = raylib_1.default.UpdateCameraWrap(camera);
-        if (newCamera) {
-            camera.position = newCamera.position;
-            camera.target = newCamera.target;
-            camera.up = newCamera.up;
-            camera.fovy = newCamera.fovy;
-            camera.type = newCamera.type;
-        }
-    } }, colors, camera, enums, maths, misc, texture, { 
+var raylib = __assign({}, raylib_1.default, colors, camera, enums, maths, misc, texture, models, shaders, { 
     // aliases
     Camera: camera.Camera3D });
 module.exports = raylib;
